requirements.txt
PyQt6>=6.0.0
python-docx>=0.8.11
openpyxl>=3.0.9
PyPDF2>=2.10.0
Pillow>=9.0.0
markdown>=3.3.6
lxml>=4.7.1 
beautifulsoup4>=4.10.0 
PyMuPDF>=1.18.0 